+++ src/tests/test_args.py
@@ -1,7 +1,7 @@
 import pytest  # type: ignore
 
 from normcap.args import create_argparser
-from normcap.gui.settings import init_settings
+from normcap.gui.settings import Settings
 
 # Specific settings for pytest
 # pylint: disable=redefined-outer-name,protected-access,unused-argument
@@ -24,8 +24,7 @@ def test_argparser_defaults_are_complete(argparser_defaults):
             "reset",
             "tray",
             "update",
-            "verbose",
-            "very_verbose",
+            "debug_level",
         ]
     )
     assert args_keys == expected_options
@@ -40,10 +39,10 @@ def test_argparser_help_is_complete():
 
 
 def test_all_argparser_attributes_in_settings(argparser_defaults):
-    settings = init_settings("normcap", "settings", initial={}, reset=False)
+    settings = Settings("normcap", "settings", init_settings={})
 
     for arg in argparser_defaults:
-        if arg in ["verbose", "very_verbose", "reset"]:
+        if arg in ["debug_level", "reset"]:
             continue
         assert arg in settings.allKeys()
 
@@ -52,6 +51,7 @@ def test_all_argparser_attributes_in_settings(argparser_defaults):
 
 
 def test_argparser_defaults_are_correct(argparser_defaults):
-    assert argparser_defaults["reset"] is False
-    assert argparser_defaults["verbose"] is False
-    assert argparser_defaults["very_verbose"] is False
+    assert argparser_defaults.pop("reset") is False
+    assert argparser_defaults.pop("debug_level") == "warning"
+    for value in argparser_defaults.values():
+        assert value is None